+++ src/lib/api/actions/submitFairLaunch.ts
@@ -366,7 +366,7 @@ async function submitToSmartContract({
     id: new BN(order.orderId),
     subMarketId: new BN(order.subMarketId),
     choiceId: new BN(order.choiceId),
-    amount: new BN(order.amount * Math.pow(10, 9)),
+    amount: new BN(BigInt(order.amount) * BigInt(Math.pow(10, 9))),
     requestedPricePerShare: 0.5,
   }));
